requirements/common.txt
dj-database-url==0.4.1
django-configurations==2.0.0
django-debug-toolbar==1.5.0
django-dotenv==1.4.1
django-extensions==1.7.2
Django==1.10.5
psycopg2==2.6.2

djangorestframework==3.4.7
Markdown==2.6.7
django-filter==0.15.2
djangorestframework_gis==0.10.1
django_rest_swagger==2.0.6

jsonschema==2.5.1
django_material==0.11.0
django-crispy-forms==1.6.1
django-leaflet==0.19.0
djangorestframework-filters==0.9.1